+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 更改日志 | CHANGE LOG
 
+## v0.9.9/2022-05-16
+
+- [v0.9.8 <=> v0.9.9](https:///github.com/Zuoqiu-Yingyi/siyuan-theme-dark-plus/compare/v0.9.8...v0.9.9)
 - 为资源块(挂件块, iframe 块, 视频块与音频块)添加`在新窗口打开资源`菜单项 | Add the `Open the resource in a New Window` menu item for resource blocks (widget blocks, iframe blocks, video blocks and audio blocks).
 - 为子窗口添加窗口置顶/取消置顶菜单项 | Add the `Topmost` menu item for subwindows.
 - 使用 <kbd>Shift + Alt + 鼠标中键</kbd> 单击文档在新窗口打开 [Monaco 编辑器](https://github.com/microsoft/monaco-editor)并编辑文档的 kramdown 源码 | Click on a document to open [Monaco Editor](https://github.com/microsoft/monaco-editor) in a new window and edit the kramdown source code of document using the <kbd>Shift + Alt + Middle Mouse Button</kbd>.